\name{EdwardsMBA}
\docType{data}
\alias{EdwardsMBA}
\title{MBA Data Set from Edwards (1994)}
\description{A real data set to illustrate single-index response surface models. Among a group of MBA students, job satisfaction is highest when preference for various kinds of work matches the extent to which they actually engage in those kinds of work.}
\usage{
data(EdwardsMBA)
}
\format{A matrix containing 172 rows (observations) and 9 columns (variables): 
\tabular{ll}{
\code{PCPRE} \tab preferences (PRE) for 'planning and coordinating' (PC) kind of work\cr
\code{DMPRE} \tab preferences (PRE) for 'decision making' (DM) kind of work\cr
\code{EIPRE} \tab preferences (PRE) for 'exchanging information' (EI) kind of work\cr
\code{MRPRE} \tab preferences (PRE) for 'motivating and rewarding others' (MR) kind of work\cr
\code{PCACT} \tab actual engagement (ACT) in 'planning and coordinating' (PC) kind of work\cr
\code{DMACT} \tab actual engagement (ACT) in 'decision making' (DM) kind of work\cr
\code{EIACT} \tab actual engagement (ACT) in 'exchanging information' (EI) kind of work\cr
\code{MRACT} \tab actual engagement (ACT) in 'motivating and rewarding others' (MR) kind of work\cr
\code{AVGSAT4} \tab a measure of job satisfaction\cr}}
\references{
Edwards JR (1994), "The study of congruence in organizational behavior research: Critique and a proposed alternative," \emph{Organizational Behavior and Human Decision Processes}, 58, pp. 51-100 (erratum, 58, pp. 323-325).  
}
\keyword{datasets}
